src/requirements.txt
git+https://github.com/CIAT-DAPA/aclimate_v3_orm
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.2
ecdsa==0.19.1
fastapi==0.115.12
greenlet==3.2.2
h11==0.16.0
idna==3.10
psycopg2==2.9.10
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
python-dotenv==1.1.0
python-jose==3.4.0
requests==2.32.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
